A first draft of the concatenating_strings receipe
There is either a bug in doc_code_merge, or I'm not using it as
I should, but at this time the tool is generating 3 copies of each
code snippet. I'm pushing this anyway, but we need to fix the problem
before the readable version of the file can be built.